+++ tests/qversion.py
@@ -1,4 +1,0 @@
-from PySide6.QtCore import qVersion
-
-s = ''  # type: str
-s = qVersion()